+++ ScoutSuite/providers/gcp/facade/memorystoreredis.py
@@ -1,9 +1,9 @@
 from ScoutSuite.core.console import print_exception
 from ScoutSuite.providers.gcp.facade.basefacade import GCPBaseFacade
 from ScoutSuite.providers.gcp.facade.utils import GCPFacadeUtils
-from ScoutSuite.providers.utils import run_concurrently
 
-class CloudMemorystoreRedisFacade(GCPBaseFacade):
+
+class MemoryStoreRedisFacade(GCPBaseFacade):
     def __init__(self):
         super().__init__('redis', 'v1beta1')
