+++ ScoutSuite/providers/gcp/resources/memorystore/base.py
@@ -1,7 +1,8 @@
 from ScoutSuite.providers.gcp.resources.projects import Projects
-from ScoutSuite.providers.gcp.resources.cloudmemorystore.redis_instances import RedisInstances
+from ScoutSuite.providers.gcp.resources.memorystore.redis_instances import RedisInstances
 
-class CloudMemorystore(Projects):
+
+class MemoryStore(Projects):
     _children = [ 
         (RedisInstances, 'redis_instances')
      ]
