+++ ScoutSuite/providers/gcp/services.py
@@ -1,7 +1,7 @@
 from ScoutSuite.providers.base.services import BaseServicesConfig
 from ScoutSuite.providers.gcp.facade.base import GCPFacade
 from ScoutSuite.providers.gcp.resources.cloudsql.base import CloudSQL
-from ScoutSuite.providers.gcp.resources.cloudmemorystore.base import CloudMemorystore
+from ScoutSuite.providers.gcp.resources.memorystore.base import MemoryStore
 from ScoutSuite.providers.gcp.resources.cloudstorage.base import CloudStorage
 from ScoutSuite.providers.gcp.resources.gce.base import ComputeEngine
 from ScoutSuite.providers.gcp.resources.iam.base import IAM
@@ -22,7 +22,7 @@ class GCPServicesConfig(BaseServicesConfig):
         facade = GCPFacade(default_project_id, project_id, folder_id, organization_id, all_projects)
 
         self.cloudsql = CloudSQL(facade)
-        self.cloudmemorystore = CloudMemorystore(facade)
+        self.cloudmemorystore = MemoryStore(facade)
         self.cloudstorage = CloudStorage(facade)
         self.computeengine = ComputeEngine(facade)
         self.iam = IAM(facade)
